refactor: move test parsing-result types to homeboy-extension-contract
Continues the test subsystem breakdown. The four test parsing-result types
(CoverageOutput, UncoveredFile, TestFailureSummaryItem, TestSummaryOutput) form
a pure self-contained closure and move to the contract crate.

The parse functions (build_test_summary, parse_test_results_file/_text,
parse_coverage_file, the From<RawTestFailure> conversion, and the ParseSpec
adapters) stay in core. Core re-exports the types so all crate::extension::test
paths stay stable.

These are field dependencies of the aggregate TestCommandOutput cluster - this
plus TestCounts/TestScopeOutput/test-analysis clears more of its field graph
toward moving the top-level test result cluster.
